latest

+++ database/migrations/2025_09_15_075124_create_product_grades_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('product_grades', function (Blueprint $table) {
             $table->id();
-            $table->decimal('grade', 4, 1);
+            $table->float('grade', 4, 1);
             $table->timestamps();
         });
     }
